+++ source/PrimeStatsUpdater.h
@@ -11,10 +11,10 @@ public:
 
    virtual bool   UpdateGroupStats(string candidateName) { return false; };
 
-   virtual bool   UpdateGroupStats(int64_t theK, int32_t theB, int32_t theN, int32_t theC) { return false; };
+   virtual bool   UpdateGroupStats(int64_t theK, int32_t theB, int32_t theN, int64_t theC) { return false; };
 
    virtual bool   InsertCandidate(string candidateName, int64_t theK, int32_t theB, int32_t theN,
-                                  int32_t theC, int32_t theD, double decimalLength) { return false; };
+                                  int64_t theC, int32_t theD, double decimalLength) { return false; };
 
    bool     UpdateStats(string   userID,
                         string   teamID,
