+++ source/StatsUpdater.h
@@ -23,7 +23,7 @@ public:
    virtual bool   RollupGroupStats(bool deleteInsert) { return false; };
 
    virtual bool   InsertCandidate(string candidateName, int64_t theK, int32_t theB, int32_t theN,
-                                  int32_t theC, int32_t theD, double decimalLength) { return false; };
+                                  int64_t theC, int32_t theD, double decimalLength) { return false; };
 
 protected:
    int32_t       ib_NeedsDoubleCheck;
