+++ source/WagstaffStatsUpdater.h
@@ -11,10 +11,10 @@ private:
 
    bool  UpdateGroupStats(string candidateName);
 
-   bool  UpdateGroupStats(int64_t theK, int32_t theB, int32_t theN, int32_t theC);
+   bool  UpdateGroupStats(int64_t theK, int32_t theB, int32_t theN, int64_t theC);
 
    bool  InsertCandidate(string candidateName, int64_t theK, int32_t theB, int32_t theN,
-                                               int32_t theC, int32_t theD, double decimalLength);
+                                               int64_t theC, int32_t theD, double decimalLength);
 };
 
 #endif // #ifndef _WagstaffStatsUpdater_
